+++ fj-doc-freemarker/src/test/java/test/org/fugerit/java/doc/freemarker/process/TestFreemarkerKotlinStep.java
@@ -2,6 +2,7 @@ package test.org.fugerit.java.doc.freemarker.process;
 
 import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.core.cfg.ConfigRuntimeException;
+import org.fugerit.java.core.lang.helpers.BooleanUtils;
 import org.fugerit.java.doc.base.kotlin.parse.DocKotlinParser;
 import org.fugerit.java.doc.base.process.DocProcessContext;
 import org.fugerit.java.doc.base.process.DocProcessData;
@@ -12,13 +13,15 @@ import org.fugerit.java.doc.json.parse.DocJsonParser;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Properties;
 
 @Slf4j
 public class TestFreemarkerKotlinStep {
 
-    private static final String KTS_PATH_SAMPLE = "coverage/kts/sample-2-coverage.kts";
+    private static final String KTS_PATH_SAMPLE = "coverage/kts/document-kotlin.kts";
 
     @Test
     public void testKotlinStep() throws Exception {
@@ -26,6 +29,7 @@ public class TestFreemarkerKotlinStep {
         DocProcessData data = new DocProcessData();
         // fail no parser
         FreeMarkerKotlinStep stepFailNoParser = new FreeMarkerKotlinStep( -1 );
+        stepFailNoParser.setCustomConfig( new Properties() );
         Assert.assertThrows( ConfigRuntimeException.class, () -> stepFailNoParser.process(context, data ) );
         // fail no kts
         FreeMarkerKotlinStep stepFailNoKtsPath = new FreeMarkerKotlinStep();
@@ -37,8 +41,10 @@ public class TestFreemarkerKotlinStep {
         // test using template-path
         customConfig.setProperty( FreeMarkerComplexProcessStep.ATT_TEMPLATE_PATH, KTS_PATH_SAMPLE );
         stepOk.setCustomConfig( customConfig );
+        List<People> listPeople = Arrays.asList(new People("Luthien", "Tinuviel", "Queen"), new People("Thorin", "Oakshield", "King"));
+        stepOk.getCustomConfig().setProperty( FreeMarkerComplexProcessStep.ATT_MAP_ATTS, "listPeople" );
         log.info( "customConfig >> : {}", stepOk.getCustomConfig() );
-        stepOk.process( context, data );
+        stepOk.process( context.withAtt( "listPeople", listPeople ), data );
         Assert.assertNotNull( data.getCurrentXmlData() );
         // test using kts-path
         customConfig.setProperty( FreeMarkerKotlinStep.ATT_KTS_PATH, KTS_PATH_SAMPLE );
@@ -50,3 +56,4 @@ public class TestFreemarkerKotlinStep {
     }
 
 }
+